Abstract Board Controller made to simplify code
