+++ src/test/java/model/cabal/internals/DrawStackTest.java
@@ -155,6 +155,4 @@ class DrawStackTest {
 
         return new DrawStack(cards);
     }
-
-
 }
